requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
asynch==0.2.4
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
ciso8601==2.3.1
clarabel==0.9.0
click==8.1.7
clickhouse-driver==0.2.7
clickhouse-sqlalchemy==0.2.3
comm==0.2.2
contourpy==1.3.0
cvxpy==1.5.3
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
dill==0.3.8
ecos==2.0.14
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.114.0
fastjsonschema==2.20.0
fonttools==4.53.1
greenlet==3.0.3
h11==0.14.0
idna==3.8
immutabledict==4.2.0
importlib_metadata==8.4.0
infi.clickhouse-orm==2.1.3
ipykernel==6.29.5
ipython==8.27.0
ipython-genutils==0.2.0
iso8601==2.1.0
jedi==0.19.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.7
leb128==1.0.8
lz4==4.3.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
minio==7.2.8
nbformat==5.10.4
nest_asyncio==1.6.0
numpy==1.26.4
optbinning==0.19.0
ortools==9.9.3963
osqp==0.6.7.post1
packaging==24.1
pandas==2.1.4
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
platformdirs==4.3.2
plotly==5.24.0
prompt_toolkit==3.0.47
protobuf==5.28.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pwlf==2.2.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.1
pydantic_core==2.23.3
pyDOE==0.3.8
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
pyzmq==25.1.2
qdldl==0.1.7.post4
referencing==0.35.1
requests==2.32.3
ropwr==1.0.0
rpds-py==0.20.0
scikit-learn==1.5.1
scipy==1.14.1
scs==3.2.7
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.51
stack-data==0.6.2
starlette==0.38.5
tenacity==9.0.0
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
uvicorn==0.30.6
wcwidth==0.2.13
wheel==0.43.0
zipp==3.20.1
zstd==1.5.5.1
